refactored the entire project to better adhere to SOLID principles and clean architecture
