+++ src/main/java/Diglott/Main.java
@@ -1,0 +1,8 @@
+// Main.java
+package Diglott;
+
+public class Main {
+    public static void main(String[] args) {
+        new AppLauncher().start();
+    }
+}
