+++ src/main/java/UI/components/FileSelector.java
@@ -1,19 +1,19 @@
-package UI;
+package UI.components;
 
 import javax.swing.*;
 import java.io.File;
 
 public class FileSelector {
 
-    public File selectBookFile() {
+    public static File selectBookFile() {
         JFileChooser chooser = new JFileChooser();
-        chooser.setDialogTitle("Choose a book file (.txt or .pdf)");
+        chooser.setDialogTitle("Choose a book file (.txt or .pdf or .epub)");
 
         int result = chooser.showOpenDialog(null);
         if (result == JFileChooser.APPROVE_OPTION) {
             File file = chooser.getSelectedFile();
             String name = file.getName().toLowerCase();
-            if (name.endsWith(".txt") || name.endsWith(".pdf")) {
+            if (name.endsWith(".txt") || name.endsWith(".pdf") || name.endsWith(".epub")) {
                 return file;
             }
         }
