+++ src/main/java/application/usecase/TranslatePageUseCase.java
@@ -1,0 +1,8 @@
+// TranslatePageUseCase.java
+package application.usecase;
+
+import domain.model.Page;
+
+public interface TranslatePageUseCase {
+    void execute(Page page);
+}
