+++ src/main/java/domain/gateway/Translator.java
@@ -1,0 +1,6 @@
+package domain.gateway;
+
+// Translator.java
+public interface Translator {
+    void addWord(String word) throws Exception;
+}
