+++ src/main/java/domain/gateway/WordTransliterator.java
@@ -1,0 +1,6 @@
+package domain.gateway;
+
+// Transliterator.java
+public interface WordTransliterator {
+    String transliterate(String input);
+}
