+++ src/main/java/domain/model/Page.java
@@ -1,4 +1,4 @@
-package Book;
+package domain.model;
 
 import java.util.ArrayList;
 import java.util.Collections;
