+++ src/main/java/infrastructure/importer/PdfBookImporter.java
@@ -1,5 +1,6 @@
-package Book;
+package infrastructure.importer;
 
+import domain.gateway.BookImporter;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.text.PDFTextStripper;
