+++ src/main/java/infrastructure/translation/TransliterationHandler.java
@@ -1,8 +1,9 @@
-package Translation;
+package infrastructure.translation;
 
 import com.ibm.icu.text.Transliterator;
+import domain.gateway.WordTransliterator;
 
-public class TransliterationHandler {
+public class TransliterationHandler implements WordTransliterator {
 
     private final Transliterator transliterator;
 
@@ -10,6 +11,7 @@ public class TransliterationHandler {
         this.transliterator = Transliterator.getInstance("Any-Latin; Latin-ASCII");
     }
 
+    @Override
     public String transliterate(String input) {
         if (input == null) return "";
         return transliterator.transliterate(input);
